Go back to using cast for adding user message
We already get the event when the message is added. I think I would rather allow
the agent to continue to have messages added; they will just be moved to the end
of the process queue if the thread is not in a state that allows adding messages
(i.e. run in progress)