Require last_message_id when connecting to an agent (#13)
This pull request adds a new requirement for the `last_message_id`
parameter when connecting to an agent. It must always be explicitly set,
even if it is set to `nil`. This ensures that client code is proactively
setting the value, preventing situations where an old value is used and
later messages are not displayed.