Update assistant_id on each connection
It doesn't make any sense to require that this value be specified on
connect but then ignore it if the agent is already running. Let's put
the responsibility of ensuring that any change to an assistant ID is
appropriate on the client application rather than forcing the client
application to jump through unnecessary hoops to update the value
separately from the connection request.