Return error tuple when operating on dead process
Most of the functions of the GptAgent module expect to receive a pid of
a running GptAgent process as their first argument. Prior to this
change, if that pid did not represent a running process (e.g. the
GptAgent process has been shut down for any reason), then the function
call would result in an unhandled (by this library) exception. This
changes the behavior such that we return an error tuple that the calling
code can more easily handle.