Retry runs on time/token based rate limiting
When the run status indicates that OpenAI has rate-limited the request
due to tokens per minute or requests per minute, we will retry the
requests after a configurable delay and up to a configurable number of
times (defaults to no retries).

With this commit, we slightly modify the error code returned with the
RunFailed event to indicate whether the failure is due to temporary rate
limiting or quota being reached as well as whether a retry will be
attempted. We return the event even when a retry will be attempted, so
that the subscribers can be aware that rate-limiting is occuring.